Small tweak
Hey,

Just a small tweak to make sure that 'unique' mailbox names are passed through to 'Get-MailboxFolderStatistics'.  I just found this script yesterday and it has saved me a ton of time from having to write this from scratch, so THANK YOU.  I ran into an issue with it error on mailbox names not being unique, hence this tiny edit.

Thanks again.